Andrew/organize test errors (#254)
* Separate into separate files

* Separate into separate files

* Add timeouts to error log

* Add state to profiler

* Fix bug in detecting plotly's internal js

* Add plotly to dev deps

* Add basic tests for page generator

* Add changelog
