+++ src/py/pyproject.toml
@@ -48,6 +48,7 @@ dev = [
   "async-timeout",
   "mypy>=1.14.1",
   "poethepoet>=0.30.0",
+  "plotly>=6.0.0",
 ]
 
 #docs = [
